% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_funcs.R
\name{agg_abund}
\alias{agg_abund}
\title{Aggregate abundances based on a factor variable}
\usage{
agg_abund(x, agg.var)
}
\arguments{
\item{x}{A data set containing columns '\code{agg.var}', '\code{N_predict}'
and '\code{N_real}'. The last two are obtained from a calls to
\code{agTrendTMB::sample_N} and \code{agTrendTMB::get_real}.}

\item{agg.var}{Variable used for aggregation}
}
\description{
Aggregate abundances based on a factor variable
}
