% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTrendTMB-package.R
\docType{package}
\name{agTrendTMB-package}
\alias{agTrendTMB}
\alias{agTrendTMB-package}
\title{agTrendTMB: Interpolate Survey Counts For Site-Level Aggregating}
\description{
Uses time time trends and constrined Gaussian processes to interpolate abundance counts for aggregated analysis.
}
\author{
\strong{Maintainer}: Devin Johnson \email{devin.johnson@noaa.gov}

}
\keyword{internal}
