% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_agTrend_ssl_mgcv.R
\name{extract_tweedie}
\alias{extract_tweedie}
\title{Tweedie distribution properties in fitted model}
\usage{
extract_tweedie(object)
}
\arguments{
\item{object}{A model object from \code{mgcv::gam} fitted with \code{family=tw()}}
}
\value{
a list with Tweedie parameters \code{p}, \code{phi}
}
\description{
Obtain the Tweedie distribution properties from a fitted model object
}
