% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_agTrend_ssl_mgcv.R
\name{fit_agTrend_ssl_mgcv}
\alias{fit_agTrend_ssl_mgcv}
\title{Site-level fitting function}
\usage{
fit_agTrend_ssl_mgcv(.x, obl.corr = FALSE, debug = FALSE)
}
\arguments{
\item{.x}{Data frame for an individual site. Must have been passed through \cite{prep_for_fit} first!}

\item{obl.corr}{Logical. Should correction be made for oblique photos.}

\item{debug}{Logical. Step into fitting function for interactive evaluation}

\item{model}{Model for interpolation: 'const', 'lin', or 'gp' (see Details)}
}
\value{
A list with the following elements:
\itemize{
\item summary Parameter and abundance estimate information?
\item raw_data Original data plus oblique photo correction estimates
\item q_output quality control diagnostic data
}
}
\description{
Fit separate discrete-normal distribution models to an Steller sea lion counts at a single site
}
\details{
There are 3 models which can be used for within site interpolation:
\itemize{
\item \code{const} This interpolates using a simple mean and variance for the site
\item \code{lin} A linear fit with time is used for interpolation
\item \code{gp} i.e., Gaussian Process. This model used a time trend plus a random walk of order 2 (RW(2)) to interpolate
counts. Further, the RW(2) process can be further restrained to minimize 'wiggliness' of the curve when necessary. This
prevents overfitting of the curve.
}
}
