% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_growth.R
\name{get_growth}
\alias{get_growth}
\title{Calculate average yearly growth}
\usage{
get_growth(x, years)
}
\arguments{
\item{x}{An anumdance sample from \cite{sample_N} or \cite{agg_abund}}

\item{years}{A 2-vector that contains the first year and the last of the desired trend}
}
\description{
Calculate average yearly growth
}
