% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_real.R
\name{get_real}
\alias{get_real}
\title{Add realized SSL survey values to the prediction sample}
\usage{
get_real(fit_list, N_sample)
}
\arguments{
\item{fit_list}{A list object from \code{fit_agtrend_ssl}}

\item{N_sample}{A matrix of abundance samples from \code{sample_N}}
}
\description{
Add realized SSL survey values to the prediction sample
}
