% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot A Site-Level Fit For Diagnostic Checking}
\usage{
plot_fit(fit_list, title = NULL, plot = TRUE)
}
\arguments{
\item{fit_list}{A list resulting from a call to \code{fit_agTrend_ssl}}

\item{title}{plot title}

\item{plot}{Logical. Whether or not to produce the plot or just return the \code{ggplot2} object for further modification}
}
\value{
A \code{ggplot2} object
}
\description{
Plot A Site-Level Fit For Diagnostic Checking
}
