% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_fit.R
\name{prep_for_fit}
\alias{prep_for_fit}
\title{Trim sites that do not fit the criteria for model fitting}
\usage{
prep_for_fit(
  x,
  site.name = "SITE",
  abund.name = "counts",
  time.name = "year",
  model.cuts = c(3, 10),
  non.zero = 2,
  total = 10
)
}
\arguments{
\item{x}{Steller sea lion count data for one DPS and age class, e.g., eDPS nonpups}

\item{site.name}{Name of column with site names}

\item{abund.name}{Name of the column with the abundance counts}

\item{time.name}{Name of the column with the time (e.g., year) data}

\item{model.cuts}{Number of nonzero surveys for each model type (see Details)}

\item{non.zero}{Minimum number of nonzero counts for inclusion in model fitting. Must be >2.}

\item{total}{Minimum total number of sea lions observed at the site over the time span considered}
}
\description{
Trim sites that do not fit the criteria for model fitting
}
\details{
The \code{model.cuts} is a 2-vector argument that provides the boundaries for each model type, 'const' (constant meand term),
'lin' (linear slope mean term) or 'gp' (Gaussian process mean term). E.g., the default value is
\code{model.cuts=c(3,10)} which means that any site with less than or equal to 3 nonzero
survey values with be fit with a 'const' model and if the number of nonzero counts is in (3,10] a
'lin' model is fitted, and so on.
}
