% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_N.R
\name{sample_N}
\alias{sample_N}
\title{Sample abundance from fitted agTrendTMB model}
\usage{
sample_N(fit_list, size = 1000)
}
\arguments{
\item{fit_list}{A list object resulting from \code{fit_agTrend_ssl}}

\item{size}{Sample size. Defaults to 1000}
}
\description{
Sample abundance from fitted agTrendTMB model
}
