% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_N_mgcv.R
\name{sample_N_mgcv}
\alias{sample_N_mgcv}
\title{Sample abundance from fitted agTrendTMB model}
\usage{
sample_N_mgcv(x, size = 1000)
}
\arguments{
\item{x}{A list object resulting from \code{fit_agTrend_ssl}}

\item{size}{Sample size. Defaults to 1000}
}
\description{
Sample abundance from fitted agTrendTMB model
}
