% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_funcs.R
\name{summary_agg}
\alias{summary_agg}
\title{Summarize aggregation results}
\usage{
summary_agg(x, ci.prob = 0.95)
}
\arguments{
\item{x}{An aggregated abundance object from \code{agg_abund}}

\item{ci.prob}{Confidence (credible) interval value}
}
\description{
Summarize aggregation results
}
